Fixes Sound Preference bug
When the notification sound is in the external storage, the Settings
can't display a name for the sound. This could have been solved by
adding the READ_EXTERNAL_STORAGE permission, but I think that it's
better to not add unnecesary permissions and just display a static
text under the Settings. Other application do the same (i.e. QKSMS)